+++ src/Smartstore.Core/Platform/Widgets/Extensions/ViewContextExtensions.cs
@@ -1,8 +1,9 @@
-﻿using Microsoft.AspNetCore.Mvc.Rendering;
+﻿using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.AspNetCore.Mvc.ViewFeatures;
 using Microsoft.Extensions.Options;
 
-namespace Smartstore.Web.Rendering
+namespace Smartstore.Core.Widgets
 {
     public static class ViewContextExtensions
     {